backend/requirements.txt
# Django
Django>=4.2,<5.0
django-cors-headers>=4.3.0
django-filter>=23.5
Pillow>=10.0.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Google Sheets Integration
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
python-dotenv>=1.0.0
